requirements.txt
flask==2.0.1
flask-cors==3.0.10
Pillow
torch
torchvision
numpy
gunicorn==20.1.0
